Moved components around so that they are displayed in the stock details component
