+++ frontend/src/components/LandingPage/Weather/Weather.js
@@ -2,9 +2,9 @@ import React from 'react';
 import './Weather.css'
 
 const weatherData = {
-    city: "Allen" + " ",
-    temperature: "65" + '\xB0F ',
-    picture: "/partly cloudy.png"
+    city: "Allen ",
+    temperature: "65 ",
+    picture: "/partly_cloudy.png"
 };
 
 
@@ -14,7 +14,7 @@ function Weather() {
             <span>
                <b>{city}</b>
                {temperature}
-               <img className="medium-icon" src={picture} />
+               <img className="medium-icon" src={picture} alt="weather"/>
             </span>
         );
 }
