+++ frontend/src/components/LandingPage/Weather/Weather.test.js
@@ -1,6 +1,6 @@
-const weatherData = {
-    city: "allen",
-    temperature: "65"
-  };
+// const weatherData = {
+//     city: "allen",
+//     temperature: "65"
+//   };
 
   // render an object like this {weatherData}
