+++ frontend/src/components/StockDetails/StockQuote/StockQuote.test.js
